handlers: renamed credentials -> taskmetadata
renamed package credentials as taskmetadata. Also, moved
credentials fetching methods to a file of theier own.
Lastly, made some exported const's to be not exported
anymore.
